refactor: modularize scanning utilities
